external module for hbci mapping used for camt and mt940
