+++ onlinebanking-bankinggateway/src/main/java/de/adorsys/multibanking/parsing/TransactionsParser.java
@@ -3,8 +3,8 @@ package de.adorsys.multibanking.parsing;
 import de.adorsys.multibanking.domain.BalancesReport;
 import de.adorsys.multibanking.domain.Booking;
 import de.adorsys.multibanking.domain.response.LoadBookingsResponse;
-import de.adorsys.multibanking.hbci.model.HbciObjectMapper;
-import de.adorsys.multibanking.hbci.model.HbciObjectMapperImpl;
+import de.adorsys.multibanking.hbci.mapper.HbciObjectMapper;
+import de.adorsys.multibanking.hbci.mapper.HbciObjectMapperImpl;
 import org.apache.commons.io.IOUtils;
 import org.kapott.hbci.GV.parsers.ISEPAParser;
 import org.kapott.hbci.GV.parsers.SEPAParserFactory;
