--- onlinebanking-bankinggateway/src/test/java/de/adorsys/multibanking/Test.java
@@ -1,7 +1,0 @@
-package de.adorsys.multibanking;
-
-import org.junit.Ignore;
-
-@Ignore
-public class Test {
-}
