+++ onlinebanking-hbci4java/pom.xml
@@ -22,6 +22,11 @@
             <version>5.3.4-SNAPSHOT</version>
         </dependency>
         <dependency>
+            <groupId>de.adorsys.multibanking</groupId>
+            <artifactId>onlinebanking-mapper</artifactId>
+            <version>5.3.4-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
         </dependency>
