+++ onlinebanking-hbci4java/src/main/java/de/adorsys/multibanking/hbci/job/NewStandingOrderJob.java
@@ -20,6 +20,7 @@ import de.adorsys.multibanking.domain.request.TransactionRequest;
 import de.adorsys.multibanking.domain.response.EmptyResponse;
 import de.adorsys.multibanking.domain.transaction.AbstractScaTransaction;
 import de.adorsys.multibanking.domain.transaction.StandingOrder;
+import de.adorsys.multibanking.hbci.model.HbciCycleMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.kapott.hbci.GV.AbstractHBCIJob;
@@ -35,7 +36,7 @@ import java.util.List;
 
 @RequiredArgsConstructor
 @Slf4j
-public class NewStandingOrderJob extends ScaRequiredJob<StandingOrder, EmptyResponse> {
+public class NewStandingOrderJob extends ScaRequiredJob<StandingOrder, EmptyResponse> implements HbciCycleMapper {
 
     private final TransactionRequest<StandingOrder> transactionRequest;
 
@@ -62,10 +63,10 @@ public class NewStandingOrderJob extends ScaRequiredJob<StandingOrder, EmptyResp
             gvDauerSEPANew.setParam("firstdate", standingOrder.getFirstExecutionDate().toString());
         }
         if (standingOrder.getCycle() != null) {
-            gvDauerSEPANew.setParam("timeunit", hbciObjectMapper.cycleToTimeunit(standingOrder.getCycle())); // M
+            gvDauerSEPANew.setParam("timeunit", cycleToTimeunit(standingOrder.getCycle())); // M
             // month, W
             // week
-            gvDauerSEPANew.setParam("turnus", hbciObjectMapper.cycleToTurnus(standingOrder.getCycle())); // 1W = every
+            gvDauerSEPANew.setParam("turnus", cycleToTurnus(standingOrder.getCycle())); // 1W = every
             // week, 2M = every two months
         }
         gvDauerSEPANew.setParam("execday", standingOrder.getExecutionDay()); // W: 1-7, M: 1-31
