+++ onlinebanking-hbci4java/src/main/java/de/adorsys/multibanking/hbci/job/ScaRequiredJob.java
@@ -26,6 +26,8 @@ import de.adorsys.multibanking.domain.response.AbstractResponse;
 import de.adorsys.multibanking.domain.response.AuthorisationCodeResponse;
 import de.adorsys.multibanking.domain.response.UpdateAuthResponse;
 import de.adorsys.multibanking.domain.transaction.AbstractScaTransaction;
+import de.adorsys.multibanking.hbci.mapper.HbciObjectMapper;
+import de.adorsys.multibanking.hbci.mapper.HbciObjectMapperImpl;
 import de.adorsys.multibanking.hbci.model.*;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -61,6 +63,7 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 @Slf4j
 public abstract class ScaRequiredJob<T extends AbstractScaTransaction, R extends AbstractResponse> {
 
+    private static HbciDialogRequestMapper hbciDialogRequestMapper = new HbciDialogRequestMapperImpl();
     static HbciObjectMapper hbciObjectMapper = new HbciObjectMapperImpl();
 
     private HbciTanSubmit hbciTanSubmit = new HbciTanSubmit();
@@ -355,7 +358,7 @@ public abstract class ScaRequiredJob<T extends AbstractScaTransaction, R extends
     }
 
     private HbciDialogRequest createDialogRequest(HBCICallback hbciCallback) {
-        return hbciObjectMapper.toHbciDialogRequest(getTransactionRequest(), hbciCallback);
+        return hbciDialogRequestMapper.toHbciDialogRequest(getTransactionRequest(), hbciCallback);
     }
 
     private HbciConsent getConsent() {
