+++ onlinebanking-hbci4java/src/main/java/de/adorsys/multibanking/hbci/model/HbciDialogRequestMapper.java
@@ -1,0 +1,10 @@
+package de.adorsys.multibanking.hbci.model;
+
+import de.adorsys.multibanking.domain.request.AbstractRequest;
+import org.kapott.hbci.callback.HBCICallback;
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface HbciDialogRequestMapper {
+    HbciDialogRequest toHbciDialogRequest(AbstractRequest transactionRequest, HBCICallback callback);
+}
