+++ pom.xml
@@ -63,6 +63,7 @@
         <module>multibanking-persistence-mongodb</module>
         <module>multibanking-persistence-jpa</module>
         <module>onlinebanking-facade</module>
+        <module>onlinebanking-mapper</module>
         <module>onlinebanking-figo</module>
         <module>onlinebanking-finapi</module>
         <module>onlinebanking-hbci4java</module>
